+++ src/clj/grenada/exporters.clj
@@ -1,13 +1,15 @@
 (ns grenada.exporters
-  "
+  "Procedures for writing collections of Things to disk in various formats.
 
   See also grenada.exporters.pretty."
   (:require [clojure.java.io :as io]
             [darkestperu.jar :as jar]
             [plumbing.core :as plumbing :refer [safe-get]]
+            [schema.core :as s]
             grimoire.util
             [grenada
              [config :refer [config]]
+             [schemas :as schemas]
              [things :as t]
              [utils :as gr-utils]]
             [grenada.utils.jar :as gr-jar]))
@@ -29,43 +31,77 @@
 
 ;;;; Printing primitives
 
-(defn prn-spit [path x]
+(defn prn-spit
+  "Like clj::clojure.core/spit, but uses clj::clojure.core/prn for printing
+  instead of whatever spit uses."
+  [path x]
   (with-open [writer (io/writer path)]
     (binding [*out* writer] (prn x))))
 
 
 ;;;; Hierarchical filesystem exporter
 
-(defn- exp-map-fs-hier [m out-dir]
+(defn- exp-map-fs-hier
+  "Writes Thing M to its proper place below OUT-DIR."
+  [m out-dir]
   (let [data-dir (io/file out-dir (coords->path (safe-get m :coords)))
         data-path (io/file data-dir (safe-get config :datafile-name))]
     (io/make-parents data-path)
     (prn-spit data-path m)
     data-path))
 
-(defn fs-hier [data out-dir]
+(defn fs-hier
+  "Writes a collection of Things to their appropriate places below OUT-DIR.
+
+  Read them back like this:
+
+  ````````clojure
+  (->> out-dir
+       grenada.utils/ordinary-file-seq
+       (map slurp)
+       (map grenada.reading/read-string))
+  ````````"
+  [data out-dir]
   (doseq [m data]
     (exp-map-fs-hier m out-dir)))
 
 
 ;;;; Flat filesystem exporters
 
-(defn fs-flat [data out-file]
+(defn fs-flat
+  "Writes a collection of Things to OUT-FILE.
+
+  Read them back like this:
+  ```````clojure
+  (->> out-file
+       slurp
+       grenada.reading/read-string)
+  ```````
+
+  See also clj::grenada.exporters.pretty/pprint-fs-flat unless you're working in
+  pre-1.7.0 Clojure."
+  [data out-file]
   (prn-spit out-file data))
 
 
 ;;;; JAR exporter
 
-(defn- thing->jar-entry [t]
+(defn- thing->jar-entry
+  "Returns the entry for T for the entries-map to be given to
+  clj::grenada.utils.jar/jar-from-entries-map."
+  [t]
   [(gr-utils/str-file
      (safe-get config :jar-root-dir)
      (coords->path (safe-get t :coords))
      (safe-get config :datafile-name))
    (jar/->string-entry (prn-str t))])
 
-;; TODO: Document the feature that one can pass additional entries for the POM
-;;       file in coords-out. (RM 2015-08-07)
-(defn jar [things out-dir coords-out]
+(s/defn ^:always-validate jar
+  "Creates a Datadoc JAR containing the THINGS and a pom.xml for this JAR in
+  OUT-DIR.
+
+  The JAR's filename and the pom.xml will be derived from COORDS-OUT."
+  [things out-dir coords-out :- schemas/JarCoordsWithDescr]
   (gr-jar/jar-from-entries-map (->> things
                                     (map thing->jar-entry)
                                     (into {}))