+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.0-rc.3
+
+ - [BREAKING] Remove `grenada.utils/safe-select-keys`.
+ - Add lots of doc strings and some schemas for arguments and return values.
+ - Add optional arguments to grenada.reading/read-string.
+
 ## 1.0.0-rc.2
 
  - Remove `grenada.mergers`. - This hadn't been working with the new model